moving to float species

+++ lucene/core/src/java/org/apache/lucene/internal/vectorization/DefaultVectorUtilSupport.java
@@ -381,11 +381,11 @@ final class DefaultVectorUtilSupport implements VectorUtilSupport {
 
   @Override
   public int filterByScore(
-      int[] docBuffer, double[] scoreBuffer, double minScoreInclusive, int upTo) {
+      int[] docBuffer, float[] scoreBuffer, float minScoreInclusive, int upTo) {
     int newSize = 0;
     for (int i = 0; i < upTo; ++i) {
       int doc = docBuffer[i];
-      double score = scoreBuffer[i];
+      float score = scoreBuffer[i];
       docBuffer[newSize] = doc;
       scoreBuffer[newSize] = score;
       if (score >= minScoreInclusive) {
